+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/Multiselect/Multiselect.tsx
@@ -1,3 +1,4 @@
+//@ts-nocheck
 import { UIElement } from '@/domains/collection-flow';
 import {
   Chip,