+++ packages/ui/src/components/organisms/Form/DynamicForm/hooks/internal/useEvents/types/index.ts
@@ -1,0 +1,18 @@
+import { ICommonFieldParams } from '../../../../types';
+
+import { IFormElement } from '../../../../types';
+
+export type TElementEvent =
+  | 'onChange'
+  | 'onMount'
+  | 'onBlur'
+  | 'onFocus'
+  | 'onSubmit'
+  | 'onClick'
+  | 'onUnmount';
+
+export interface IFormEventElement<TElements extends string, TParams = ICommonFieldParams>
+  extends IFormElement<TElements, TParams> {
+  formattedValueDestination: string;
+  formattedId: string;
+}