+++ packages/ui/src/components/organisms/Form/Validator/utils/validate/helpers.ts
@@ -1,0 +1,6 @@
+import jsonLogic from 'json-logic-js';
+import { IValidationRule } from '../../types';
+
+export const isShouldApplyValidation = (rule: IValidationRule, context: object) => {
+  return Boolean(jsonLogic.apply(rule.value, context));
+};